requirements.txt
fastapi
uvicorn
torch
einops
numpy
pillow
pytorch-lightning
opencv-python
